Refactor CategoryRepository and CategoryService

+++ src/main/java/aamm/order/service/CategoryService.java
@@ -1,12 +1,14 @@
 package aamm.order.service;
 
 import java.util.HashMap;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import aamm.order.Repository.CategoryRepository;
 import aamm.order.model.Category;
+import aamm.order.model.Product;
 
 @Service
 public class CategoryService {
@@ -14,23 +16,53 @@ public class CategoryService {
     CategoryRepository categoryRepository;
 
     public boolean addCategory(Category category) {
-        return categoryRepository.add(category);
+        if (CategoryRepository.isCategoryExists(category.getSlug())) {
+            return false;
+        }
+        categoryRepository.add(category);
+        return true;
     }
 
     public Category getCategory(String slug) {
-        return categoryRepository.getCategory(slug);
+        if (CategoryRepository.isCategoryExists(slug)) {
+            return categoryRepository.getCategory(slug);
+        } else {
+            return null;
+        }
     }
 
-    public void updateCategory(String slug, Category category) {
-        categoryRepository.updateCategory(slug, category);
+    public boolean updateCategory(String slug, Category category) {
+        if (CategoryRepository.isCategoryExists(slug)) {
+            categoryRepository.updateCategory(slug, category);
+            return true;
+        } else {
+            return false;
+        } 
     }
 
-    public void deleteCategory(String slug) {
-        categoryRepository.deleteCategory(slug);
+    public boolean deleteCategory(String slug) {
+        if (CategoryRepository.isCategoryExists(slug)) {
+            categoryRepository.deleteCategory(slug);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public HashMap<String, Category> getCategories() {
-        return categoryRepository.getCategories();
+        if (categoryRepository.getCategories() != null) {
+            return categoryRepository.getCategories();
+        } else {
+            return null;
+        }
+    }
+
+    public HashMap<String, Product> getCategoryProducts(String slug) {
+        if (CategoryRepository.isCategoryExists(slug)) {
+            return categoryRepository.getProducts(slug);
+        } else {
+            return null;
+        }
     }
 
 }
